Split import mapping funcs into new file
